rejig settings to be singleton class
